Refactor code for ufs preparation
Move methods to ufs_internal.go
Move struct UFSInitResult to ufs_internal.go
Rewrite logic of UFS initialization
